chore(API): refactor property controller and add validation middleware create validation middleware [Starts#167161258]
