+++ server/routes/propertyRoutes.js
@@ -1,5 +1,6 @@
 import express from 'express';
 import multiparty from 'connect-multiparty';
+import ValidationMiddleware from '../middlewares/validation'
 import PropertyController from '../controllers/propertyController';
 
 const multipartyMiddle = multiparty();
@@ -8,7 +9,7 @@ const router = express.Router();
 
 router.get('/properties', PropertyController.viewAllProperties);
 router.get('/property/:id', PropertyController.viewPropertyById);
-router.post('/property', multipartyMiddle, PropertyController.postNewProperty);
+router.post('/property', multipartyMiddle, ValidationMiddleware.createPropertyValidation, PropertyController.postNewProperty);
 router.delete('/property/:id', PropertyController.deleteProperty);
 router.patch('/property/:id', multipartyMiddle, PropertyController.updatePropertyDetails);
 router.patch('/property/:id/sold', PropertyController.markAsSold);
